refactor: Make comment as float just like go ast
