+++ field.go
@@ -1,7 +1,6 @@
 package gg
 
 import (
-	"fmt"
 	"io"
 )
 
@@ -46,23 +45,4 @@ func (f *ifield) render(w io.Writer) {
 	}
 	writeString(w, f.separator)
 	f.value.render(w)
-}
-
-func parseNodes(in []interface{}) []Node {
-	ns := make([]Node, 0, len(in))
-	for _, v := range in {
-		ns = append(ns, parseNode(v))
-	}
-	return ns
-}
-
-func parseNode(in interface{}) Node {
-	switch v := in.(type) {
-	case Node:
-		return v
-	case string:
-		return String(v)
-	default:
-		panic(fmt.Errorf("invalid input: %s", v))
-	}
 }
