+++ node.go
@@ -1,6 +1,9 @@
 package gg
 
-import "io"
+import (
+	"fmt"
+	"io"
+)
 
 type Node interface {
 	render(w io.Writer)
@@ -9,4 +12,28 @@ type Node interface {
 // Embed accept a close clause to build a node.
 func Embed(fn func() Node) Node {
 	return fn()
+}
+
+// parseNodes will parse valid input into node slices.
+func parseNodes(in []interface{}) []Node {
+	ns := make([]Node, 0, len(in))
+	for _, v := range in {
+		ns = append(ns, parseNode(v))
+	}
+	return ns
+}
+
+// parseNode will parse a valid input into a node.
+// For now, we only support two types:
+// - Native Node
+// - golang string
+func parseNode(in interface{}) Node {
+	switch v := in.(type) {
+	case Node:
+		return v
+	case string:
+		return String(v)
+	default:
+		panic(fmt.Errorf("invalid input: %s", v))
+	}
 }
